Declare several packet fields `pub(crate)`
This makes them visit through the crate, so we can just take ownership of their values when converting from packets to packet groups. There's a cool example of ownership going on here, especially on the `String` and `Vec<u8>` types since they'll still be immutable through all the passing around.