Modify `packets` to use `anyhow`
I think this sort of gets `packets` so that it uses `anyhow` error handling.

Unfortunately it wasn't clear how to test the new `anyhow::Error`s, so this mad the tests worse, although presumably made the user experience better. I left TODOs on some of the places where this can/should be addressed.